1. change unint_ptr to unsigned long long 2. delete empty lines

+++ gasonwrapper/gason.hpp
@@ -39,8 +39,8 @@ union JsonValue {
     JsonValue(double x)
     : fval(x) {
     }
-    JsonValue(JsonTag tag = JSON_NULL, void *payload = nullptr) {
-        assert((uintptr_t)payload <= JSON_VALUE_PAYLOAD_MASK);
+    JsonValue(JsonTag tag = JSON_NULL, void *payload = nullptr) {        
+        assert((unsigned long long)payload <= JSON_VALUE_PAYLOAD_MASK);
 //        assert((uintptr_t)payload <= JSON_VALUE_PAYLOAD_MASK);
         ival = JSON_VALUE_NAN_MASK | ((uint64_t)tag << JSON_VALUE_TAG_SHIFT) | (uintptr_t)payload;
     }
